re-implementing 'rescan' using mem-mapped blocks
This code is pretty simple and a great playing ground for creating a
good API for Tx::Iterator

This shows that I can now iterate through all blocks with exacly
one pass only, reading all transactions in the chain in
less than 20 minutes (probably half that if it was on an SSD).

Additionally, this makes it trivial to make the multithreaded
code be safe and make all the actual reading mutex-free.