Activate Nov 2019 & May 2020 HFs by height instead of MTP
and update the definitions of the Mainnet, Testnet (v3), and Regtest chain params accordingly.  The heights for all 3 chains were taken by verifying the values in both BU and BCHN code bases.

NOTE: For TestNet v3, there is a 1 block discrepancy between BU (1378461) and BCHN (1378460) for the height of the May 2020 hardfork.  I arbitrarily selected the value from BCHN, I'm not sure which is correct.