Take action on a static code checker finding an issue.
[code-quality]
We got a warning about a dereference of a nullptr. But that implied we
have an impossible combination of member variables.
So, lets assert on that being impossible so;
a) we get a fail-fast if this ever happens.
b) the static code checker will stop warning us.

Notice that asserts are compiled out of release executables. So this is
effectively a no-op for releases.