Fix lifetime of the diskSpaceChecker class
As the DiskSpaceChecker owns a deadlinetimer, which depends on the asio
io_service, we need to ensure it is destructed before the io_service is.

The io_service, and now the diskSpaceChecker as well, are specifically
shut down, in order.
Application is still a singleton that dies after main() ends, but will
have nearly no members at that time.